+++ src/routes/AppRouter.tsx
@@ -9,7 +9,6 @@ import appRoutes, { Route } from './routes';
 import ScrollToTop from './ScrollToTop';
 import { useStore } from '../store/storeContext';
 import NotFound404 from '../views/404';
-import RouteTracker from '../components/AnalyticsRouteTracker';
 import PageWrapper from './PageWrapper';
 
 interface Props {
@@ -27,7 +26,6 @@ const AppRouter: FC<Props> = observer(({ routes = appRoutes }) => {
   return (
     <Router>
       <ScrollToTop />
-      <RouteTracker />
       <Switch>
         {publicRoutes.map(route => (
           <RouterRoute